move program validator to new file
